refactor(profile/reports): use parallel routes instead of server side sections
