--- app/profile/[id]/reports/sections/index.ts
@@ -1,1 +1,0 @@
-export * from './listening-days-section'
